tests can run again
